Crate twitch_client

Overview

Rust Twitch Client is a library for the Twitch REST API written in Rust!

It uses hyper with native_tls as https client and serde for the serialization and deserialization of the REST requests and responses.

Examples

use twitch_client::*;

fn main() {
    let twitch_client = TwitchClient::new("<YOUR_TWITCH_CLIENT_ID>").unwrap();

    match twitch_client.top_games(TopGamesParams::default()) {
        Ok(top_games) => {
            println!("Total games: {}", top_games.total());
            println!("---");
            for game_info in top_games.top() {
                println!("Game: {}, Viewers: {}", game_info.game().name(), game_info.viewers());
            }
            println!("---");
        },
        Err(err) => println!("Failed to retrieve top games: {}", err),
    }
}

Reexports

pub use param::*;

Modules

error
Error and Result module.

model
Twitch return types.

param
Parameters for TwitchClient methods.

Structs

TwitchClient
Readonly client for the Twitch REST API.