Module twitch_client::model::stream

Twitch streams.

Streams are video broadcasts that are currently live. They have a broadcaster and are part of a channel.

Reexports

pub use model::DateString;

pub use model::UrlString;

pub use model::image::ImageLinks;

pub use model::channel::Channel;

Structs

ChannelStream
Twitch stream information of a specific channel.

FeaturedStream
Featured (promoted) stream.

FeaturedStreams
Featured (promoted) streams.

Stream
Basic Stream type.

Streams
Streams that are queried by a number of parameters sorted by number of viewers descending.

StreamsSummary
Summary of current streams.