Trait twitch_client::model::TwitchLinks

pub trait TwitchLinks {
    fn links(&self) -> &BTreeMap<String, String>;

    fn get_expected_link(&self, link_key: &str) -> &String { ... }
}

A Twitch Model that contains http links to the called resource and other related Twitch resources.

Required Methods

fn links(&self) -> &BTreeMap<String, String>

Returns a map with all http links of the Twitch result.

Provided Methods

fn get_expected_link(&self, link_key: &str) -> &String

Returns the http link corresponding to the given key.

The key is expected to exist in the links() map. The safe way to return a link is to use the BTreeMap.get(key) method.

Panics

Panics if the given key does not exist.