Struct twitch_client::TwitchClient

pub struct TwitchClient { /* fields omitted */ }

Readonly client for the Twitch REST API.

Currently Twitch API version 3 is used.

By using the Twitch Client you agree to follow the Twitch Developer Services Agreement and the Twitch Terms of Service. This library is in no way affiliated with, authorized, maintained, sponsored or endorsed by Twitch or any of its affiliates or subsidiaries

Examples

use twitch_client::*;

let twitch_client = TwitchClient::new("<YOUR_TWITCH_CLIENT_ID>").unwrap();

match twitch_client.top_games(TopGamesParams::default()) {
    Ok(top_games) => println!("Total games: {}", top_games.total()),
    Err(err) => println!("Failed to retrieve top games: {}", err),
}

Methods

impl TwitchClient

fn new<S: Into<String>>(client_id: S) -> Result<TwitchClient>

Constructs a new client instance with a new hyper https client using native tls.

Since 2016-08-06 a Twitch Client ID is required. Instructions for obtaining it can be found at the Twitch API Documentation.

fn with_hyper_client<S: Into<String>>(client_id: S,
hyper_client: Client)
-> TwitchClient

Constructs a new client instance using the provided hyper client.

Note that the provided hyper client needs to use a tls connection.

Since 2016-08-06 a Twitch Client ID is required. Instructions for obtaining it can be found at the Twitch API Documentation.

impl TwitchClient

fn top_games(&self, params: TopGamesParams) -> Result<TopGames>

Get games by number of viewers.

Returns a list of games objects sorted by number of current viewers on Twitch, most popular first.

fn ingests(&self) -> Result<Ingests>

Get list of ingests.

Returns a list of ingest objects.

fn basic_info(&self) -> Result<BasicInfo>

Get top level links object and authorization status.

Basic information about the API and authentication status. If you are authenticated, the response includes the status of your token and links to other related resources.

fn stream(&self, channel: &str) -> Result<ChannelStream>

Get stream object.

Returns a stream object if live.

fn streams(&self, params: StreamsParams) -> Result<Streams>

Get stream object.

Returns a list of stream objects that are queried by a number of parameters sorted by number of viewers descending.

fn featured_streams(&self,
params: FeaturedStreamsParams)
-> Result<FeaturedStreams>

Get a list of featured streams.

Returns a list of featured (promoted) stream objects.

fn streams_summary(&self,
params: StreamsSummaryParams)
-> Result<StreamsSummary>

Get a summary of streams.

Returns a summary of current streams.

fn channel(&self, channel: &str) -> Result<Channel>

Get channel object.

Returns a channel object.