Struct twitch_client::TwitchClient

pub struct TwitchClient {
    // some fields omitted
}

Readonly client for the Twitch API.

Currently Twitch API version 3 is used.

By using the Twitch Client you agree to follow the Twitch API Terms of Service and the Twitch Terms of Service. This library is in no way affiliated with, authorized, maintained, sponsored or endorsed by Twitch or any of its affiliates or subsidiaries

Examples

use twitch_client::*;

let twitch_client = TwitchClient::new();
let top_games = twitch_client.top_games(TopGamesParams::default()).unwrap();
println!("Total games: {}", top_games.total());

Methods

impl TwitchClient

fn new() -> TwitchClient

Constructs a new instance without client id and with a default hyper client.

It is highly recommended to specify a client id to avoid being rate limited by Twitch with the with_client_id method.

fn with_client_id(self, client_id: &str) -> TwitchClient

Sets the Twitch client id.

See https://github.com/justintv/twitch-api#rate-limits for more information.

fn with_hyper_client(self, hyper_client: Client) -> TwitchClient

Sets a custom configured hyper client.

See hyper::client::Client for more information.

impl TwitchClient

fn top_games(&self, params: TopGamesParams) -> Result<TopGames>

Get games by number of viewers.

Returns a list of games objects sorted by number of current viewers on Twitch, most popular first.

fn ingests(&self) -> Result<Ingests>

Get list of ingests.

Returns a list of ingest objects.

fn basic_info(&self) -> Result<BasicInfo>

Get top level links object and authorization status.

Basic information about the API and authentication status. If you are authenticated, the response includes the status of your token and links to other related resources.

fn stream(&self, channel: &str) -> Result<ChannelStream>

Get stream object.

Returns a stream object if live.

fn streams(&self, params: StreamsParams) -> Result<Streams>

Get stream object.

Returns a list of stream objects that are queried by a number of parameters sorted by number of viewers descending.

fn featured_streams(&self, params: FeaturedStreamsParams) -> Result<FeaturedStreams>

Get a list of featured streams.

Returns a list of featured (promoted) stream objects.

fn streams_summary(&self, params: StreamsSummaryParams) -> Result<StreamsSummary>

Get a summary of streams.

Returns a summary of current streams.

fn channel(&self, channel: &str) -> Result<Channel>

Get channel object.

Returns a channel object.